Lots of new documentation for heat
